chore: Add MoviesListScreen
- Extract MoviesListScreen from MainActivity to a dedicated file
